Refactor in main and created new files for interface and cases
