+++ interface.rb
@@ -1,0 +1,16 @@
+class Interface
+  puts '--------------------------------------------------'
+  puts 'Welcome to my library! ͡° ͜ʖ ͡°'
+  def interface
+    puts '--------------------------------------------------'
+    puts 'Select an option: '
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals'
+    puts '7 - Exit'
+    puts '--------------------------------------------------'
+  end
+end
